% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_gtin}
\alias{data_gtin}
\title{dataset - GTIN}
\description{
Conjunto de dados com alguns GTIN para testar algumas funções
}
\usage{
seidatalab::data_gtin
}
\format{
[,1] gtin_17
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
data_gtin
}
\keyword{datasets}

