% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ncm}
\alias{data_ncm}
\title{dataset - NCM}
\description{
Conjunto de dados com alguns NCM para testar algumas funções
}
\usage{
seidatalab::data_ncm
}
\format{
[,1] ncm_17
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
data_ncm
}
\keyword{datasets}

