% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncm_valido}
\alias{ncm_valido}
\title{dataset - NCM 2022}
\description{
Conjunto de dados com os códigos da Nomenclatura Comum do MERCOSUL vigentes.
}
\usage{
seidatalab::ncm_valido
}
\format{
[,1] ncm
}
\author{
Jonatas Silva do Espirito Santo
}
\source{
ncm_valido
}
\keyword{datasets}
